Textual description of the pizza done on customer side.
This commit moves all the functionality of textual describing a pizza
individual from the server to the client application. Also, it changes
how this textual description is displayed by showing (on a very
problematic way) it on a thought balloon (Actually what it does is
displaying the message using the 'cowthink' program and cropping the cow
away.

Signed-off-by: hugocpolos <hugo.cpolos@gmail.com>